Reorder preprocessing steps to avoid false trimming of existential subqueries [#150988530]
Orca remove outer references from order grouping columns in GbAgg. Orca also
trims an existential subquery whose inner expression is a GbAgg  with no grouping
columns by replacing it with a Boolean constant. However in same caes
applying the first preprocessing step affects the latter and produces an
unintended trimming of existential subqueries.

This commit changes the order of these two preprocessing step to avoid
that complications.

E.g
```
SELECT * from foo where not exits (SELECT sum(bar.a) from bar where
foo.a = bar.a GROUP BY foo.a;
```

In this example the grouping column is an outer reference and it is
removed by `PexprRemoveSuperfluousOuterRefs`. And the next preprocessing
step `PexprTrimExistentialSubqueries` sees that the GbAgg has no
grouping colums and implys `NOT EXISTS` as `false`.

Therefore we change the order and this fixes the problem.

Bump version to 2.46.3