Fix filename validation when restoring pre-content-id backups
The glob syntax used to validate filenames for segment dump filenames
in the "old" backup format is only correct for single-digit dbids, as
it assumes the use of regex syntax that glob doesn't actually support.

This commit modifies get_filename_for_content to glob for all dump
files present and then explicitly match the expected filenames with a
regular expression.

It also modifies generate_filename to account for files created on a
standby master, which was not correctly handled previously.