Add gp_request_fts_probe_scan() UDF.
Using this UDF, external entities can request / trigger FTS probe instead of
waiting for it to get triggered at regular interval. One usecase currently for
it is after starting walrep mirror, to trigger the probe to reflect mirror-up on
primary instead of waiting for a minute (default fts interval) to get the chnage
reflected on primary.

Signed-off-by: Xin Zhang <xzhang@pivotal.io>