resgroup: workaround a race condition on slot reuse.
A slot maybe reused on QE while its still using by another QE which
might causes runtime error. This issue exists for a long time in theory,
but the fail rate is enlarged with a current commit 8005f559.

To workaround the issue we change the free slot pool from a stack into a
queue, slots are allocated from the head and freed to the tail, so a
slot will not be reused immediately after being freed.

In worst case all the in-use slots are not freed in time on QEs, so we
also doubled the slot pool size to have enough free slots for new
queries even in such a worst case.

In any sense this is not a fix, just a quick workaround, we will keep
working on this race condition issue. Once a root fix is ready this
workaround will be replaced.