Add xlogging for AO truncate.
Vacuum truncates AO/CO files if a insert transaction has been
aborted. The truncate for walrep needs to be communicated to mirror
segmnets via xlog. For filerep this was communicated via filrep
messages, hence AO/CO truncate was not xlogged. So, adding to xlog
truncate operation for walreplication and its only replayed in standby
mode.

Signed-off-by: Xin Zhang <xzhang@pivotal.io>
Signed-off-by: Taylor Vesely <tvesely@pivotal.io>