Temporary work-around to fix TINC storage test failures in CI
The new CI environment seems to behave in a peculiar way.  There is a
delay of several seconds between a child process dies with abort() and
the parent process receives a signal notifying child's death.  The
storage tests frequently inject PANIC faults and run select queries in
a loop until success to determine if the postmaster has finished reset
(shutdown followed by startup).  Due to the above mentioned delay, the
tests incrrectly determine that postmaster is up when the postmaster
hasn't even initiated shutdown sequence.  This commit causes a test to
wait for 30 seconds before starting to bombard the postmater with
select queries to determine if it's up.  A proper fix is in the
making, until then, this fix should make the CI green.