Add support for "order none" directive to atmsort.
This allows overriding the heuristic on whether a query has an ORDER BY.

Use the directive in one of the queries in the 'gporca' test, which
contains a subquery with an ORDER BY that fools the atmsort's usual
heuristic. The overall order of the query is not well-defined, even though
there is an ORDER BY in the subquery. The current implementation of
DISTINCT in fact always also sorts the output, which is why this test
is passing, but that is about to be relaxed soon, when we merge upstream
commit 63247bec28.