Make validation more robust for reindex TINC tests
To validate that a reindex operation indeed created a new index, the tests need
to inspect relfilenode values of the index's entry in pg_class.  Previously,
the GUC gp_select_invisible was turned on to get relfilenode prior to reindex.
Use of the GUC made the tests fickle.  E.g. running an unrelated test that
creates and drops a table with identical name as one of the index names in
reindex test is suffice to make the reindex tests fail.

Found several redundant tests, removed them on the way.