Fix gpperfmon dynamic memory usage bug
Stat sender backend is used by gpperfmon to record dynamic memory usage
of segments, previously stat sender always report the memory usage as
zero because it's vmem tracker is not initialized. segmentVmemChunks is
shared by all backend, special backends like stat sender need to read it,
so it's unnecessary to check vmemTrackerInited for it.

This commit also removes the confused mem_alloc_max.