Remove superfluous tests on built-in functions.
These tests executed various built-in functions, in different contexts
like in a FROM clause, or in a subquery. We have similar tests for
functions with different attributes in the main test suite's
'qp_functions_in_*' tests. The fact that these are built-in functions
doesn't make any difference to planning, so remove these tests as
redundant.