Removes redundant mpp6979 test
This test checks that ALTER TABLE EXCHANGE PARTITION with a table from a different
schema will successfully exchange the partition, as well as exchange the namespace
that the table is in. This test is already in src/test/regress/sql/partition.sql
and, therefore, does not need to be repeated as part of the TINC test suite.