A bunch of fixes for writable external table issues.
* Add a sanity check in formatter_import() for invalid lengths.

* In fileam.c, set the length-field of the HeapTupleHeader passed as Datum
correctly. Otherwise, if ExecInsert passes a tuple that came directly from
a table, and was not materialized in the tupletable slot, it will not have
the length set correctly.

* In ExecInsert, since external_insert() apparently takes the liberty to
modify the tuple it's passed, make sure we make a copy of it. Otherwise,
we might scribble over a tuple pointer that points directly at an on-disk
buffer.

* In externalgettup_custom(), don't go into an infinite loop if we reach
end-of-file prematurely, while the custom formatter says that it still
needs more data.

I bumped into these while trying to merge 8.4 commits. I'm not sure what
changed, but in the contrib/extprotocol regression tests, external_insert()
was now being passed a tuple that was already in HeapTuple format in the
tuple slot. That resulted in the infinite loop, because the custom formatter
read the length of the tuple incorrectly, and expected more data even though
EOF had already been reached. I'm not sure why that doesn't currently happen
on 'master', but I suspect it might be possible to devise a test case to
trigger it even there.