Add support for deparsing of DMLActionExpr.
Avoids the unpleasent "ERROR: unrecognized node type" error, when you try
to EXPLAIN a query that contains one.

I bumped into this while trying to move the "mpp21622" TINC test from the
QP-optimizer-function test suite to the main installcheck-world test suite.
The original test had the EXPLAIN commands in ignore-blocks, which is why
the test hasn't failed. But clearly it would be good to have some test
coverage for this, so I no ignore-block in the new test. It means that we
will have to update the expected output if there are plan changes to that
query, but the query is very simple so there aren't very many ways it
could be planned, so that doesn't seem like too much trouble.

Backpatch to 5X_STABLE.