Add a configuration option to disable enforcements of constraints.
Add a new EnforceConstraints hint configuration option. If it's set,
ORCA will not add Assert nodes to enforce CHECK and NOT NULL constraints
on INSERT and UPDATE statements.

This is useful for GPDB, which is prepared to enforce the constraints on
its own. In theory, the optimizer could prove the inserted row to always
satisfy the constraints, in which case it could optimize away the checks,
whereas the executor can't do that on its own. But ORCA doesn't currently
attempt to do that. The reason I'd like to enforce the constraints in
GPDB side instead of as Assertion nodes, is that you get a different error
message from the assertion node. That's annoying, because we have to
maintain alternative expected output files for tests that hit CHECK or
NOT NULL constraints because of that.

Bump version number to 2.48.0, since this is an incompatible change.