Refactor to use AOCS totalbyte code
We had code for getting the total bytes consumed by an AOCS table,
as well as code implementing the very same asking for a common
function. Extend GetAOCSTotalBytes() to deal with uncompressed or
compressed data and refactor callsites to use this function.

This also removes the need for a memory allocation in the codepaths
which only want to know the number of bytes.