Move MPP21090 tests from TINC to the main test suite.
I'm not convinced we really need to all these different combinations. But
let's at least move the test into the main test suite, so that it's easier
to work with.

These did reveal the bug fixed in commit c159ec7223 (github issue #3577),
so there is at least some value in these tests.