Change test to make the output  more stable.
The order that these INSERTs would attempt to insert the rows was not
well-defined, so it would vary from one run to another, which partition's
check constraint it hit first. That changes the error message, causing
sporadic failures. To fix, change the query so that all the matching rows
that are then being inserted go to the same partition.