Remove unstable test UPDATE.
This test UPDATE violates two constraints, and it's arbitrary which one it
fails on first. That's problematic, because the error message depends on
which constraint happens to be checked first.

We have tests for testing each failure individually, so testing a case that
fails them both doesn't seem very interesting. We will never reach the
actualy failure of both at the same time, because we bail out on first
error. So to make the output stable, remove the offending UPDATE.

I recently moved this query from TINC. It produced stable output before
commit ec3370b, because the old error message with ORCA listed both violated
constraints, but now the error only complaints about the one it happened to
check first. This happens to be working on the master pipeline currently, but
we got a failure from this on the 8.4 merge branch we're working on. I'm not
sure what changed there, perhaps a concurrent vacuum or some other timing
difference, but in any case the output is not very well-defined.