Remove insert test to validate duplicate row constraint
Similar to https://github.com/greenplum-db/gpdb/pull/3610.
This test may produce inconsistent result based on the partition which
it pick to validate the constraint. We have other places where we test
the similar intentions, so removing this flaky test.

Also removed the numbering of tests as it becomes cumbersome to maintain
the numbering.