Move CTE related tests out of TINC
Move CTE tests out of TINC and add (the ones that are not already
present) to main test suite.  optimizer_functional_part1 runs 58 CTE
tests which can be categorized as :

- cte_queries_<1-24>
Almost all these tests (with one exception as described next) are
already in the main test suite in `qp_with_clause.sql` which means we
had duplicate tests. Hence this commit removes them from TINC.

Two queries in `cte_queries24.sql`, query#6 and query#11 were missing in
`qp_with-clause.sql`. This commit adds them.

- cte_functest_<58-60>
In the old commit 629f7e7, these tests were not
moved since they produced different results on different invocations and
were marked as skipped tests. They indeed have different behavior with
Planner and ORCA. ORCA does not support updating rows when there are
multiple matches in the join condition and erros out. Planner on the
other hand allows this but the results are non-deterministic.  There is
already coverage for this error in qp_dml_joins.sql and
DML_over_joins.sql.  Hence this commit drops these tests and does not
move to main test suite.

- enable_cte_plan_space
This is a pure ORCA test which tests the number of plan alternatives
produced by ORCA when optimizer_cte_inlining is ON and OFF. This will be
moved in ORCA test suite

- icg_cte_with_values
The test is already present in `with.sql`.

- cte_functest_<22-23>_inlining_<enabled/disabled>
Moved these tests to `qp_with_functional.sql`. qp_with_functional is
tested with both inlining and noinlining.

This commit thus removes the `cte` test folder from TINC