Fix CTAS in UTILITY MODE after QueryDispatchDesc change
In the below referenced commit, we added a new boolean
validate_reloptions to the QueryDispatchDesc. In UTILITY MODE, the
query descriptor does not have its QueryDispatchDesc initialized so
the segment would PANIC. To fix this, we default validate_reloptions
to true if the QueryDispatchDesc is not initialized as expected before
the change.

Reference:
https://github.com/greenplum-db/gpdb/commit/a3c07830258d62ae86ce9418b84ff0ca2e73e514

Authors: Jimmy Yih and Abhijit Subramanya