Add gen_profile_only option to analyzedb.
For AO tables, users do not always want to run ANALYZE on a table when the
analyzedb command is run. For example, when they have already ANALYZE'ed the
table.

The --gen_profile_only option saves the modification count of the specified AO
table (or all AO tables if none is specified) so that a subsequent analyzedb
command will not ANALYZE the AO table if the modification count for the table
has not changed from the saved value.

Signed-off-by: Marbin Tan <mtan@pivotal.io>
Signed-off-by: Shoaib Lari <slari@pivotal.io>