Coverity fixes for FTS.
CID 178114: Performance inefficiencies  (PASS_BY_VALUE)
The `ProbeConnectionInfo` parameter was being passed to the
`probeSegmentHelper()` function by value. This is very inefficient. Fixed it by
passing the parameter as a reference using a pointer.

CID 178115: Calling "pqGetInt" without checking return value (as is done
elsewhere 52 out of 57 times).
The return value of `pqGetInt()` function was not being checked in the
`processProbeResponse()` function. Fixed it by checking for the return value.
If EOF is returned, a message is logged and false is returned to the caller.

Signed-off-by: Taylor Vesely <tvesely@pivotal.io>