Support exchange sub partition with external table
For example:

    alter tableA alter partition partiton1
    exchange partition partition1_subpartition1 with table external_table;

partition1 is the first level partition of tableA.
partition1_subpartition1 is a sub partition of partition1.

1) The routine ATPExecPartExchange in tablecmds.c searches
the target partition's table through parameters following
the key 'alter'. Then no matter what level the partition is, the
routine exchanges partition with the new table in the same way.
So exchanging sub partition with external table works well now.

2) Queries against partitioned tables that are altered to
use an external table as a leaf child partition fall back
to the legacy query optimizer. By calling routine
rel_has_external_partition in cdbpartition.c, GPORCA knows
if the table has an external partition. However, the routine
just searchs the first level partition child and it returns
false when the sub partition is an external table. Fix it in this
commit though searching every partition in table.

Signed-off-by: Xiaoran Wang <xiwang@pivotal.io>