Fix and move test case for MPP-22599.
The test had become useless somewhere along the years. The bug was that
if ORCA fell back to the planner, then the check that you cannot update
a distribution key column with the planner would not be made, and you
could end up with incorrectly distributed rows. The test used a multi-level
partitioned table as the target, because when the test was originally
written, multi-level partitioning was not supported by ORCA. But at some
point, support for that was added, so the test no longer tested the
original bug it was written for.

Rewrite the test using a different feature that ORCA falls back on, add
comments to make it more clear what this is supposed to test so that it
won't be broken so easily again. And finally, move the test out of TINC,
into the main regression suite, which is what I was doing when I realized
that it was broken altogether.