Fix gpcheckcat error reporting of missing entries entries.
In commit 226e8867d8, I changed the shape of the result set passed to the
processMissingDuplicateEntryResult() function, removing the "exists" column.
But I failed to update the line that extracts the primary key columns from
the result set for that change. Fix.

This should fix the failures in the gpcheckcat behave tests.