Fix a resgroup performance issue.
On low end system with 1~2 cpu cores the new queries in a cold resgroup
can suffer from a high latency when the overall load is very high.

The root cause is that we used to set very high cpu priority for gpdb
cgroups, so non gpdb process are scheduled with very low priority and
high latency. GPDB processes are also affected by this because
postmaster and other auxiliary are not put into gpdb cgroups. Even for
QD and QEs they are not put into a gpdb cgroup until their transaction
is began.

To fix this we made below changes:
* put postmaster and all its children processes into the toplevel
  gpdb cgroup;
* provide a GUC to control the cgroup cpu priority for gpdb processes
  when resgroup is enabled;
* set a lower cpu priority by default;