Use string representation of segment ids in CatMissingIssue object.
In commit 226e8867d8, I changed the CatMissingIssue object to hold the
content IDs of segments where an entry is missing in a Python list, instead
of the string representation of a PostgreSQL array (e.g. "{1,2,-1}") that
was used before. That was a nice simplification, but it turns out that
there was more code that accessed the CatMissingIssue.segids field that I
missed. It would make sense to change the rest of the code, IMHO, but to
make the CI pipeline happy quickly, this commit just changes the code back
to using a string representation of a PostgreSQL array again.

This hopefully fixes the MM_gpcheckcat behave test failures.