Remove "boundary" TINC tests with 10 MB datums, improve "toast" test.
These so-called boundary tests didn't actually go very close to the max size
of a text/bytea/varchar column. The maximum of a single datum is 1 GB, and
the value used here was only 10 MB.

I was tempted to just remove the tests, but since someone felt that we need
to test largish values, let's try to keep the coverage. So to compensate
for the removal of the TINC tests, add similar test queries to the "toast"
pg_regress test. Also make one of the large columns the distribution key,
so that we test detoasting of the distribution key.

I don't think using the subquery in the INSERT and UPDATE was the point of
the test, so I replaced that with a much faster function to generate test
data.

We had tests of a varchar's length limit in the 'varchar' test already, so
that aspect of the dml_boundary_varchar test also didn't seem interesting.

This removes the last tests from the "optimizer_functional_part2" set of
tests. Remove that from the Makefile and concourse pipeline file.