Move tests on partition pruning from TINC.
We had very similar tests in the "partition_pruning" test, even with the
same test table, so that's a natural home for this too.

I left out the "alter_tab_add_pkey" tests, because they were identical to
the "alter_tab_add_uniquekey" tests, except that the index was created for
a PRIMARY KEY rather than a UNIQUE constraint. I don't see the point of
that. Furthermore, the test queries in "alter_tab_add_pkey" didn't actually
make use of the index.