Remove obsolete and broken executor tracing facility.
This #ifdef'd code would not compile, because GroupState isn't defined.
And hasn't been for as long as the git history goes. That hints that no-one
has used this facility for years, so it seems safe to remove it. You can
use a debugger or custom temporary elog()s if you need to trace a particular
query like this.