Move tests on partition pruning after DROP COLUMN from TINC.
The indexes on these tests were quite useless. In 'alter_tab_drop_col_2',
the DROP COLUMN also dropped the index, so we got no chance to use it in
any queries. In 'alter_tab_drop_col_3', ORCA doesn't know how to use the
index for the IS FALSE query. In the translation from TINC, I removed the
former case, but kept the latter one, and also added another query where
ORCA also know how to use the index.

This removes the last tests from the QP_optimizer-functional test suite,
so remove all traces of that from the Makefiles and the Concourse pipeline.