Use a Postgres latch for wakeup of main thread from interconnect RX thread.
The problem with pthread wait conditions is that there is no way to wait
for the wakeup from anothre thread, and for other events, like a socket
becomeing readable, at the same time. We currently rely on polling on the
other events, which leads to unnecessary delays. In particular, if a QE
throws an ERROR, we will wait up to 250 milliseconds before the timeout
is reached, before waking up the QD main thread to process the error.

This commit doesn't actually address that problem yet, just changes the
signaling mechanism between the RX thread and the main thread. I'll make
the changes to avoid that delay as a separate commit, for easier review.