Move copy interrupt test from Tinc to ICW
Since GPDB has a lot of downstream code in the COPY codepath, testing
that we are still able to interrupt a COPY in progress is of interest.
Move the Tinc test to ICW and in the process make it actually test the
interruption by adding enough data into the testpath that it will be
able to hit before the copy is done. The original coding only added
100 rows to be copied which always finishes well before the interrupt
fault injection manages to run.