Remove gpmon "rows in" counters.
Except at the very top, one node's output is always another node's input,
so it seems silly to have separate counters for rows in.

The only place where "rows in" was used, was in gpperfmon's calculation of
"rows skew". Change that calculation to use "rows out" instead. That's not
exactly the same thing, but seems just as good for the purpose of measuring
skew.