Remove pointless mock test.
There's a direct call to MemoryAccounting_Reset() in MemoryContextInit().
This test was merely testing that the call is indeed there. That seems
completely uninteresting to me.

I tested what happens if the call is removed. Postmaster segfaults at
startup. So if that call somehow accidentally gets removed, we will find
out quickly even without this test.