Merge conditions into corresponding Bitmap Index Probes
There was bug in the way Bitmap Index Probes were being merged.

Consider the table query below :

SELECT * FROM foo WHERE b = 2 AND c >=6 AND c <= 6;

Let's assume there is index on coloumn 'b' and 'c'. Bitmap Index Probe
on second and third condition ('c>=6' and 'c<=6') are mergeable because
they are on the same indexed coloumn. Due to the bug, Orca would merge
second Index Probe with first Index Probe condition ('b=2') instead. This lead to a wrong 'Recheck
Condition' which if selected as a filter lead to a wrong plan. This
commit fixes this bug.
Signed-off-by: Jesse Zhang <sbjesse@gmail.com>