Split bfv_partition test into two parts.
This test was simply large, and also took quite a long time to run, so it's
nice to split it up. Furthermore, only some of the tests produce different
output with ORCA. Split the test so that the tests that use EXPLAIN, or
produce different output with ORCA for some other reason, go to a new
'bfv_partition_plans' test, and the rest remain in 'bfv_partition'.