Remove gpstop -a tests from cs_pg_twophase_switch.
These tests already have coverage for gpstop -i. In case of gpstop -a tests the
fault injected is resumed, command completes and gpstop -a is performed, which
will write checkpoint and clean shut-down. So, really there is nothing to test
as recovery will be normal means not doing anything, do not need so many
combinations testing this behavior.