Automatically regenerate gppylib catalog JSON file
Instead of having to remember to manually update the gppylib JSON
file (which has frequently been forgotten), hook the generation
into the src/backend/catalog "all" target such that it's generated
automatically when needed and thus can be removed from the repo
(removing the risk of using a stale file).

Also updates the documentation and some minor comment fixes to the
process_foreign_key script.