Add gp_replica_check for heap tables in ICW.
When configured with --enable-segwalrep, the tool will be enabled only for heap
relations. It has flagged some inconsistencies for other objects such as
sequences and appendonly tables. Once the issues are fixed, the tool will be
enabled to check for all objects created by ICW. This helps to consistently
check primary and mirrors.

Also add warning messages to gp_replica_check if files could not be opened.

Signed-off-by: Ashwin Agrawal <aagrawal@pivotal.io>